Bug/SU-702: Develop and Test Approach for Loan Re-Migration (Casuísticas) (#1923)
