+++ fineract-provider/src/main/java/org/apache/fineract/portfolio/loanaccount/domain/LoanAccountDomainService.java
@@ -124,4 +124,6 @@ public interface LoanAccountDomainService {
 
     void updateRepaymentInstalmentCharge(LoanRepaymentScheduleInstallment loanRepaymentScheduleInstallment, Integer numberOfRepayment);
 
+    void cleanUpLoan(Long loanId);
+
 }
